java stuff

+++ project_after/java.vim
@@ -1,6 +1,6 @@
 let g:syntastic_java_javac_classpath = $CLASSPATH
 
-if exists('g:project_android')
+if exists('g:project_android') && !exists('g:project_scala')
   MaqueAddCommand 'rake install start', {
         \ 'name': 'install',
         \ 'copy_to_main': 1,
